feat: restructuring
