+++ src/redux/reducers/feed/feed.test.ts
@@ -1,12 +1,12 @@
-import { WsStatus } from "../../shared/types/WebSocket/WsStatus";
+import { WsStatus } from "../../../shared/types/WebSocket/WsStatus";
 import {
   FEED_WS_OPEN,
   FEED_WS_CLOSED,
   FEED_WS_ORDER,
   FEED_WS_ERROR,
   FEED_WS_INIT,
-} from "../actions/feed";
-import { Order } from "../types/dataModels";
+} from "../../actions/feed";
+import { Order } from "../../types/dataModels";
 import { feedReducer, initialState } from "./feed";
 
 describe("feed reducer", () => {
