+++ src/redux/reducers/index.ts
@@ -1,19 +1,19 @@
-import { feedReducer } from "./feed";
+import { feedReducer } from "./feed/feed";
 import { combineReducers } from "redux";
-import { ingredientsReducer } from "./ingredients";
-import { ingredientModalReducer } from "./ingredientModal";
-import { constructorReducer } from "./constructor";
-import { orderConfirmationModal } from "./orderConfirmationModal";
-import { userReducer } from "./userReducer";
-import { orderDetailsModalReducer } from "./orderDetailsModal";
-import { ordersHistoryReducer } from "./ordersHistory";
+import { ingredientsReducer } from "./ingredients/ingredients";
+import { ingredientModalReducer } from "./ingredientModal/ingredientModal";
+import { constructorReducer } from "./constructor/constructor";
+import { orderConfirmationModalReducer } from "./orderConfirmationModal/orderConfirmationModal";
+import { userReducer } from "./user/user";
+import { orderDetailsModalReducer } from "./orderDetailsModal/orderDetailsModal";
+import { ordersHistoryReducer } from "./ordersHistory/ordersHistory";
 
 export const rootReducer = combineReducers({
   ingredients: ingredientsReducer,
   orderConstructor: constructorReducer,
   ingredientModal: ingredientModalReducer,
   orderDetailsModal: orderDetailsModalReducer,
-  orderConfirmationModal: orderConfirmationModal,
+  orderConfirmationModal: orderConfirmationModalReducer,
   user: userReducer,
   feed: feedReducer,
   ordersHistory: ordersHistoryReducer,
