+++ src/shared/types/State.ts
@@ -2,10 +2,10 @@ import { IngredientModal } from "./IngredientModal";
 import { Constructor } from "./Constructor";
 import { TOrderConfirmationModal } from "./OrderConfirmationModal";
 import { Ingredients } from "./Ingredients";
-import { UserState } from "../../redux/reducers/userReducer";
+import { UserState } from "../../redux/reducers/user/user";
 import { OrdersResponse } from "../../redux/types/dataModels";
 import { rootReducer } from "../../redux/reducers";
-import { TOrderDetailsModal } from "../../redux/reducers/orderDetailsModal";
+import { TOrderDetailsModal } from "../../redux/reducers/orderDetailsModal/orderDetailsModal";
 
 export type State = {
   ingredients: Ingredients;
